requirements.txt
python-novaclient
python-keystoneclient
python-neutronclient
python-cinderclient
pyYAML
apache-libcloud
